refactor(smoke-utils): split _review.py into annotation pipeline and review-design (479 + 545 LOC)
